Move loggers into their own folder, and refactor them for more generality
