+++ robotbase/src/main/java/com/systemmeltdown/robotlog/subsystems/FeederLogger.java
@@ -1,10 +1,10 @@
-package com.systemmeltdown.robotlog.commands;
+package com.systemmeltdown.robotlog.subsystems;
 
 import com.systemmeltdown.robot.subsystems.FeederSubsystem; // Imported for javadoc (for now)
 
 /**
  * A class for logging the commands used by the {@link FeederSubsystem}.
  */
-class FeederCommandsLogger {
+class FeederLogger {
     
 }
