+++ robotbase/src/main/java/com/systemmeltdown/robotlog/subsystems/ShooterLogger.java
@@ -1,4 +1,4 @@
-package com.systemmeltdown.robotlog.commands;
+package com.systemmeltdown.robotlog.subsystems;
 
 import com.systemmeltdown.robot.subsystems.ShooterSubsystem; // Imported for javadoc (for now)
 
@@ -9,6 +9,6 @@ import com.systemmeltdown.robot.subsystems.ShooterSubsystem; // Imported for jav
  * 
  * @category Logging
  */
-class ShooterCommandsLogger {
+class ShooterLogger {
     
 }
